1.1.4(8) release

+++ SleepModel/HEMPagingFlowLayout.m
@@ -38,7 +38,7 @@
             continue;
         }
         
-        if (fabsf(attributes.center.x - proposedContentOffsetCenterX) < fabsf(candidateAttributes.center.x - proposedContentOffsetCenterX))
+        if (fabs(attributes.center.x - proposedContentOffsetCenterX) < fabs(candidateAttributes.center.x - proposedContentOffsetCenterX))
         {
             candidateAttributes = attributes;
         }
